Bug fix in list and vector eval
If a value was returned from an expression then
it would be modified by being inserted into
the list or vector. If this value was the result of a symbol
lookup then this would modify the symbol value.

let* form now takes either vectors or lists

All stage 3 tests now pass