bbc-basic: Use a two-dimensional array for Z%().
Rather than allocating a variable number of elements of a
one-dimensional array for different types of object, allocate each one
as a row of a four-column array.  This isn't as space-efficient, but
it's clearer and will make memory management a lot easier (no need to
worry about fragmentation, for instance).