bbc-basic: Rough but apparently functional garbage-collector in step 4.
It's a fairly simple mark-and-sweep job, with a slighly mad arrangement
for tracking values that _might_ be in BASIC variables.  It's full of
debugging PRINTs and assorted other unpleasantness, but it does seem to
be working correctly, not leaking heap or shadow stack even in the face
of errors, so I want to get it committed before I break it.