bbc-basic: Rework try*/catch*.
This separates out the trying part into its own function, which
returns any exception that got thrown, or returns the result if no
exception occurred.  This has two significant advantages.  First, it
means that the shadow stack gets reset before the 'catch*' clause it
run, so unrechable objects allocated by the failed code become
collectable.  Second, it makes it possible (not that I've done so) to
adjust the code to cope with ARM BBC BASIC's failure to restore LOCAL
variables before running an ON ERROR LOCAL handler.  FNtry can now
safely assume that none of _its_ LOCAL variables will be corrupted
because the ones that matter are unique to FNtry, and any FNtry called
further down the stack will not propagate or raise an exception.