vala Makefile: define an empty .SUFFIXES.
This fixes an annoying make behaviour where, if the Vala compiler
leaves around a file called foo.vala.c (as it does in some forms of
compile failure), make will automatically try to rebuild foo.vala from
it. This fails, because a single valac-output source file will
basically never compile using make's default cc command – and it's
lucky it does fail, because it would overwrite a real Vala source file
if it didn't!

Defining .SUFFIXES to the empty string suppresses all the built-in
things GNU make thinks it knows about how to compile things, and
forces it to rely only on the rules actually written down in this
Makefile.