bbc-basic: Simplify 'first' and 'rest'.
Rather than special-casing the empty list/vector, arrange that the
dedicated empty list/vector objects have the correct values in their
first and rest pointers.  Still dithering about whether to do that
to 'nil' as well.

This does mean that the internal FNfirst and FNrest don't error on an
empty list any more.  Nothing should have depended on that, but it might
mask some bugs.