bbc-basic: Simplify type codes.
Z%(x,2) and Z%(x,3) are always pointers or nil, so there's no need to
record whether they're used: FNgc_mark can just unconditionally mark
them.  This means that the only interesting question is whether Z%(x,1)
is a pointer, a string, or something else, which can be encoded in two
bits.  This still leaves us using five bits for the type code, but I
think the avoidance of special-case code is worth spending a bit on.