bbc-basic: Ensure that "Escape" errors propagate beyond mal.
Rather than printing an error message and ending the program, it's
better to turn off our error handler and then re-raise the error.  This
ensures that if, for instance, mal is started from the Supervisor in
RISC OS, the error gets propagated back to the Supervisor error handler
and displayed properly.

This required a slight adjustment to the top-level ON ERROR handler.