bbc-basic: Radically simplify string storage.
Rather than having a separately indexed array of strings, S$(), with
its own free-list in S%(), we now keep string values in an array,
Z$(), whose indices parallel those in Z%(), so Z$(x) is the string
value of the object at Z%(x,y).  This saves a fair amount of
complexity at the expense of slightly more memory usage.