bbc-basic: Fix accidental quadratic behaviour in string handling.
The code for constructing long strings in FNstring_concat and 'slurp'
was accidentally quadratic in the size of output.  This meant that it
consumed BASIC heap space in inordinate quantities for storing strings.
Mal doesn't monitor the usage of the BASIC heap, so it didn't run the
garbage collector and instead suffered a fatal error.  Making these
functions use sane amounts of memory avoids that problem.