bbc-basic: Slight tweak to heap size.
Making the mal heap very slightly smaller allows the self-hosting test
to complete on a 64-bit host.

This is really not the right approach, though.  I should really just
give in and specify "-size 1024k" to Brandy, or have the GC somehow
detect that we're wasting a lot of memory on strings and it should run
early.  At the moment, there's a tiny range of heap sizes that
actually work.