step 1: Test a hash-map with more than one entry.
All the existing tests in step 1 use empty or single-element hash-maps.
That avoids any dependency on the ordering of printed hash-maps, but it
does mean that some code paths (e.g. printing the space between hash-map
elements) aren't tested.  Add a test that creates a three-element
hash-map and uses some regexp trickery to check that all the right keys
and values come out.